1_try

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: eunjeong <eunjeong@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/23 02:54:54 by eunjeong          #+#    #+#             */
-/*   Updated: 2022/12/23 03:22:44 by eunjeong         ###   ########.fr       */
+/*   Updated: 2022/12/23 15:34:32 by eunjeong         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ int		ft_pointer(void *p);
 int		ft_putptr(size_t nbr);
 int		ft_putchar(int c);
 int		ft_puthex(unsigned int nbr, char *b);
-int		ft_putint(long int n);
+int		ft_putint(long long n);
 int		ft_putstring(char *info);
 
 size_t	ft_strlen(const char *s);
